pyproject.toml
[tool.poetry]
name = "jellyfin-music-player-qt"
version = "0.1.0"
description = ""
authors = ["felix920506 <felix920506@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.5"
python-i18n = "^0.3.9"
pyyaml = "^6.0.1"
pyqt6 = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
